Make GELF::Logger#add compatible with Logger#add
This changes the interface of `GELF::Logger#add` to be the same as
Ruby's `Logger#add`.

Before the change the `GELF::Logger` would loose messages if its `#add`
method was used like `Logger#add`.

This changes the code to use the new interface and uses the same
technique as Ruby's `Logger#add` to check for messages and facilities in
the body of `#add`.