Pre-bind sockets on creation
When sending a datagram, non-bound sockets require a DNS lookup.
Instead, bind the socket to the given addresses when creating the
socket.

Also, lazily create the socket connections, which allows the
notify methods to capture the exceptions as expected.  If the sockets
are created and bound at initialize, an exception would occur then.

Lastly, allow resetting the connections if the addresses are changed.

Fixes #31