Refactor TCP transport & support TLS
- use Rubys built-in TCPSocket or - if available - Celluloid::IO
- use more Enumerable methods for more concise code
- remove unnecessary uses of self and begin block
- fix `GELF::Transport::TCP#addresses`
- remove previously deprecated options
- extract numbers into constants
- replace `class_eval` with `define_method`

The TLS session is configured to be as secure as possible by default, using
TLS v1.2, forbidding SSLv3 and explicitly setting the cipher list to prevent
weak ciphers.

The next step will be to add support for pinning or stricter verification of
the server certificate.

Since this is a breaking change (removed deprecated options) and introduces
a major feature (TLS), the version is bumped to 3.0.0.