Initial support for TLS on TCP connections
Right now this should be considered work in progress. It has been rumored
to work but as usual YMMV.
The TLS session is configured to be as secure as possible by default, using
TLS v1.2, forbidding SSLv3 and explicitly setting the cipher list to prevent
weak ciphers.

The next step will be to add support for pinning or stricter verification of
the server certificate.